pyproject.toml
[project]
name = "cwlf-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "flask>=3.1.2",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
]
